splitting wrapper.hs into Conversion/*
