+++ src/Cooked/MockChain/Direct.hs
@@ -16,6 +16,8 @@ import Control.Monad.Except
 import Control.Monad.Identity
 import Control.Monad.Reader
 import Control.Monad.State.Strict
+import Cooked.Conversion.ToScript
+import Cooked.Conversion.ToScriptHash
 import Cooked.InitialDistribution
 import Cooked.MockChain.Balancing
 import Cooked.MockChain.BlockChain
