+++ src/Cooked/Pretty/Cooked.hs
@@ -29,6 +29,7 @@ module Cooked.Pretty.Cooked
   )
 where
 
+import Cooked.Conversion.ToScriptHash
 import Cooked.MockChain.BlockChain
 import Cooked.MockChain.GenerateTx
 import Cooked.MockChain.Staged
