+++ src/Cooked/Skeleton.hs
@@ -79,11 +79,15 @@ where
 import Cardano.Api qualified as Cardano
 import Cardano.Node.Emulator qualified as Emulator
 import Control.Monad
+import Cooked.Conversion.ToCredential
+import Cooked.Conversion.ToOutputDatum
+import Cooked.Conversion.ToPubKeyHash
+import Cooked.Conversion.ToScript
+import Cooked.Conversion.ToScriptHash
 import Cooked.Output
 import Cooked.Pretty.Class
 import Cooked.ValueUtils
 import Cooked.Wallet
-import Cooked.Wrappers
 import Data.Default
 import Data.Either.Combinators
 import Data.Function
